api/requirements.txt
django==5.1
django-cors-headers==3.10.0
djangorestframework==3.15.2
tzdata==2023.3
django-environ==0.11.2
psycopg2-binary==2.9.9
gunicorn==21.2.0
whitenoise==6.6.0
dj-rest-auth[with_social]==5.0.2
django-allauth==64.2.1
django-filter==24.3
django-colorfield==0.11.0